+++ CHANGELOG.md
@@ -35,6 +35,7 @@ You may browse a formatted and hyperlinked version of this file at <https://unpo
 - Fix smooth scrolling when prepending/appending content
 - Allow smooth scrolling when swapping a fragment
 - Introduce up.link.config.noClickableSelectors and [up-clickable=false]
+- Mute an error `AbortError: Rendering was prevented by up:location:restore listener` when up:location:restore is prevented
 
 
 3.8.0